For #5574 - Part 6: Port TabAction.Select to TabSessionInteractor and SessionControlController (#6651)
